# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    name: SJ Tech Sweden
    url: https://github.com/sj-tech-sweden/rentalcore
  description: RentalCore is a professional equipment rental management system. This
    API provides endpoints for managing jobs, customers, devices, products, rental
    equipment, and financial data.
  license:
    name: Proprietary
  title: RentalCore API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Returns a list of all products with optional filtering
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: products array
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List products
      tags:
      - products
  /products/{id}:
    get:
      description: Returns a single product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: product object
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Get a product
      tags:
      - products
  /rental-equipment:
    get:
      description: Returns rental equipment list (functionality moved to WarehouseCore)
      produces:
      - application/json
      responses:
        "410":
          description: Feature moved to WarehouseCore
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: List rental equipment
      tags:
      - rental-equipment
    post:
      consumes:
      - application/json
      description: Creates rental equipment (functionality moved to WarehouseCore)
      parameters:
      - description: Rental equipment payload
        in: body
        name: equipment
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "410":
          description: Feature moved to WarehouseCore
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Create rental equipment
      tags:
      - rental-equipment
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    description: Session cookie obtained after login
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"